made changes to lib file and moved untitest files
